+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/DispatcherServlet.java
@@ -3,8 +3,8 @@ package com.interface21.webmvc.servlet.mvc;
 import com.interface21.webmvc.servlet.ModelAndView;
 import com.interface21.webmvc.servlet.NoSuchHandlerException;
 import com.interface21.webmvc.servlet.View;
-import com.interface21.webmvc.servlet.mvc.tobe.HandlerAdapters;
-import com.interface21.webmvc.servlet.mvc.tobe.RequestHandlerMappings;
+import com.interface21.webmvc.servlet.mvc.adapter.HandlerAdapters;
+import com.interface21.webmvc.servlet.mvc.mapping.RequestHandlerMappings;
 import com.interface21.webmvc.servlet.view.JspView;
 import jakarta.servlet.ServletException;
 import jakarta.servlet.http.HttpServlet;
