+++ mvc/src/test/java/com/interface21/webmvc/servlet/mvc/SimpleHandlerAdapterTest.java
@@ -1,4 +1,4 @@
-package com.interface21.webmvc.servlet.mvc.tobe;
+package com.interface21.webmvc.servlet.mvc;
 
 import static org.assertj.core.api.Assertions.assertThat;
