+++ mvc/src/test/java/samples/SimpleController.java
@@ -1,6 +1,5 @@
 package samples;
 
-import com.interface21.webmvc.servlet.mvc.asis.Controller;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import org.slf4j.Logger;
